# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-Native Data Platform API",
    "description": "Governed local data platform: authentication and RBAC, multi-engine query gateway, DataHub-backed catalog and lineage, versioned metrics, and checkpointed evidence-bound analyses.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Users",
        "operationId": "list_users_api_v1_admin_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create User",
        "operationId": "create_user_api_v1_admin_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/roles": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Update User Roles",
        "operationId": "update_user_roles_api_v1_admin_users__user_id__roles_put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRolesUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/roles": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Roles",
        "operationId": "list_roles_api_v1_admin_roles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleResponse"
                  },
                  "type": "array",
                  "title": "Response List Roles Api V1 Admin Roles Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Role",
        "operationId": "create_role_api_v1_admin_roles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/grants": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Grants",
        "operationId": "list_grants_api_v1_admin_grants_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Grant",
        "operationId": "create_grant_api_v1_admin_grants_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/grants/{grant_id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete Grant",
        "operationId": "delete_grant_api_v1_admin_grants__grant_id__delete",
        "parameters": [
          {
            "name": "grant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Grant Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/datasources/{datasource_id}/catalog-refresh": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Catalog Refresh",
        "operationId": "catalog_refresh_api_v1_admin_datasources__datasource_id__catalog_refresh_post",
        "parameters": [
          {
            "name": "datasource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Datasource Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogRefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogRefreshResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/permission-requests/{request_id}/mock-approve": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Mock Approve Permission Request",
        "operationId": "mock_approve_permission_request_api_v1_admin_permission_requests__request_id__mock_approve_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CatalogRefreshRequest": {
        "properties": {
          "schemas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Schemas"
          },
          "secure_views": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Secure Views",
            "description": "View names the administrator has verified as safe to expose (no sensitive columns)."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CatalogRefreshRequest"
      },
      "CatalogRefreshResponse": {
        "properties": {
          "datasource_id": {
            "type": "string",
            "format": "uuid",
            "title": "Datasource Id"
          },
          "registered": {
            "type": "integer",
            "title": "Registered"
          },
          "updated": {
            "type": "integer",
            "title": "Updated"
          },
          "deactivated": {
            "type": "integer",
            "title": "Deactivated"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DatasetSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "skipped": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Skipped"
          }
        },
        "type": "object",
        "required": [
          "datasource_id",
          "registered",
          "updated",
          "deactivated",
          "items"
        ],
        "title": "CatalogRefreshResponse"
      },
...
      "GrantCreate": {
        "properties": {
          "role_id": {
            "type": "string",
            "format": "uuid",
            "title": "Role Id"
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          },
          "action": {
            "type": "string",
            "enum": [
              "discover",
              "query"
            ],
            "title": "Action"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "role_id",
          "dataset_id",
          "action"
        ],
        "title": "GrantCreate"
      },
      "GrantListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GrantResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "GrantListResponse"
      },
      "GrantResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "role_id": {
            "type": "string",
            "format": "uuid",
            "title": "Role Id"
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role_id",
          "dataset_id",
          "action",
          "created_by"
        ],
        "title": "GrantResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoleCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_-]*$",
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "RoleCreate"
      },
      "RoleResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "capabilities"
        ],
        "title": "RoleResponse"
      },
...
      "UserCreate": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._-]+$",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 12,
            "title": "Password"
          },
          "role_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Role Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "UserCreate"
      },
      "UserListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "UserListResponse"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/RoleResponse"
            },
            "type": "array",
            "title": "Roles"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "active",
          "roles",
          "created_at"
        ],
        "title": "UserResponse"
      },
      "UserRolesUpdate": {
        "properties": {
          "role_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Role Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "role_ids"
        ],
        "title": "UserRolesUpdate"
      },
...
    }
  },
...
}